+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel/gateway_selector.rs
@@ -1,10 +1,15 @@
 // Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use nym_gateway_directory::{EntryPoint, ExitPoint, Gateway, GatewayCacheHandle, GatewayType};
+use nym_gateway_directory::{
+    EntryPoint, ExitPoint, Gateway, GatewayCacheHandle, GatewayList, GatewayType,
+};
 use nym_vpn_api_client::types::ScoreThresholds;
 
-use crate::{GatewayDirectoryError, tunnel_state_machine::TunnelType};
+use crate::{
+    GatewayDirectoryError,
+    tunnel_state_machine::{TunnelSettings, TunnelType},
+};
 
 // First gateways with performance >= 75% are selected
 // Performance threshold expressed as percentage from 0 to 100
@@ -22,15 +27,15 @@ pub struct SelectedGateways {
 
 pub async fn select_gateways(
     gateway_cache_handle: GatewayCacheHandle,
-    tunnel_type: TunnelType,
-    entry_point: Box<EntryPoint>,
-    exit_point: Box<ExitPoint>,
+    tunnel_settings: &TunnelSettings,
     wg_score_thresholds: Option<ScoreThresholds>,
     mix_score_thresholds: Option<ScoreThresholds>,
 ) -> Result<SelectedGateways, GatewayDirectoryError> {
     // The set of exit gateways is smaller than the set of entry gateways, so we start by selecting
     // the exit gateway and then filter out the exit gateway from the set of entry gateways.
 
+    let entry_point = tunnel_settings.entry_point.as_ref();
+    let exit_point = tunnel_settings.exit_point.as_ref();
     if let (
         EntryPoint::Gateway {
             identity: entry_identity,
@@ -38,7 +43,7 @@ pub async fn select_gateways(
         ExitPoint::Gateway {
             identity: exit_identity,
         },
-    ) = (entry_point.as_ref(), &exit_point.as_ref())
+    ) = (entry_point, &exit_point)
         && entry_identity == exit_identity
     {
         return Err(GatewayDirectoryError::SameEntryAndExitGateway {
@@ -46,13 +51,26 @@ pub async fn select_gateways(
         });
     };
 
-    let (mut entry_gateways, exit_gateways) = match tunnel_type {
+    let (mut entry_gateways, exit_gateways) = match tunnel_settings.tunnel_type {
         TunnelType::Wireguard => {
             let all_gateways = gateway_cache_handle
                 .lookup_gateways(GatewayType::Wg)
                 .await
                 .map_err(GatewayDirectoryError::LookupGateways)?;
-            (all_gateways.clone(), all_gateways)
+
+            let entry_gateways = if tunnel_settings.bridges_enabled() {
+                GatewayList::new(
+                    all_gateways
+                        .clone()
+                        .into_iter()
+                        .filter(|gw| gw.bridge_params.is_some())
+                        .collect(),
+                )
+            } else {
+                all_gateways.clone()
+            };
+
+            (entry_gateways, all_gateways)
         }
         TunnelType::Mixnet => {
             // Setup the gateway that we will use as the exit point
@@ -72,14 +90,17 @@ pub async fn select_gateways(
     tracing::info!("Found {} entry gateways", entry_gateways.len());
     tracing::info!("Found {} exit gateways", exit_gateways.len());
 
-    let (min_wg_performance, min_mixnet_performance) =
-        high_performance_tier(tunnel_type, wg_score_thresholds, mix_score_thresholds);
+    let (min_wg_performance, min_mixnet_performance) = high_performance_tier(
+        tunnel_settings.tunnel_type,
+        wg_score_thresholds,
+        mix_score_thresholds,
+    );
     let exit_gateway = exit_point
         .lookup_gateway(&exit_gateways, min_wg_performance, min_mixnet_performance)
         .or_else(|err| {
             // When no gateways could be found, lower performance tier and try again
             if err.is_unmatched_non_specific_gateway() {
-                let (min_wg_performance, min_mixnet_performance) = medium_performance_tier(tunnel_type, wg_score_thresholds, mix_score_thresholds);
+                let (min_wg_performance, min_mixnet_performance) = medium_performance_tier(tunnel_settings.tunnel_type, wg_score_thresholds, mix_score_thresholds);
                 tracing::debug!("Could not locate high quality exit gateway. Lowering performance filter to medium and trying again");
 
                 exit_point.lookup_gateway(
@@ -107,7 +128,7 @@ pub async fn select_gateways(
         .or_else(|err| {
             // When no gateways could be found, lower performance tier and try again
             if err.is_unmatched_non_specific_gateway() {
-                let (min_wg_performance, min_mixnet_performance) = medium_performance_tier(tunnel_type, wg_score_thresholds, mix_score_thresholds);
+                let (min_wg_performance, min_mixnet_performance) = medium_performance_tier(tunnel_settings.tunnel_type, wg_score_thresholds, mix_score_thresholds);
                 tracing::debug!("Could not locate high quality entry gateway. Lowering performance filter to medium and trying again");
 
                 entry_point.lookup_gateway(