+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel/wireguard/two_hop_config.rs
@@ -25,7 +25,7 @@ pub const ENTRY_MTU: u16 = if cfg!(any(target_os = "ios", target_os = "android")
 pub const EXIT_MTU: u16 = if cfg!(any(target_os = "ios", target_os = "android")) {
     MIN_IPV6_MTU
 } else {
-    ETHERNET_V2_MTU - WG_TUNNEL_OVERHEAD * 2
+    ETHERNET_V2_MTU - (WG_TUNNEL_OVERHEAD * 2)
 };
 
 /// A struct that holds all configuration needed to setup the tunnels, tun device and forwarder.